Consolidate types into one file to match src structure
