--- types/recoil/lib/core/RecoilRoot.d.ts
@@ -1,12 +1,0 @@
-/// <reference types="react" />
-
-import { RecoilState } from './recoilValue';
-
-export interface RecoilRootProps {
-    initializeState?: (options: {
-        set: <T>(recoilVal: RecoilState<T>, newVal: T) => void;
-        setUnvalidatedAtomValues: (atomMap: Map<string, unknown>) => void;
-    }) => void;
-}
-
-export const RecoilRoot: React.FC<RecoilRootProps>;
