--- types/recoil/lib/core/node.d.ts
@@ -1,1 +1,0 @@
-export class DefaultValue {}
